sample/AgenticTranscoder/examples/router/pyproject.toml
[project]
name = "router-agent"
version = "0.1.0"
description = "Router Agent - multi-agent routing example (classify → cook/meteorologist)"
requires-python = ">=3.12"
dependencies = [
    "agentic-flow",
    "openai-agents>=0.3.2",
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "python-dotenv>=1.0",
    "chatkit",
    "httpx>=0.27",
]

[tool.uv.sources]
agentic-flow = { path = "../../../..", editable = true }

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = [".", "../../../../src"]
testpaths = ["tests"]
